requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
face-recognition==1.3.0
dlib==19.24.2
numpy==1.26.4
Pillow==10.2.0
supabase==2.3.4
pydantic==2.6.1
